+++ internal/js_ast/js_ast.go
@@ -501,6 +501,11 @@ type EClass struct{ Class Class }
 type EIdentifier struct {
 	Ref Ref
 
+	// If we're inside a "with" statement, this identifier may be a property
+	// access. In that case it would be incorrect to remove this identifier since
+	// the property access may be a getter or setter with side effects.
+	MustKeepDueToWithStmt bool
+
 	// If true, this identifier is known to not have a side effect (i.e. to not
 	// throw an exception) when referenced. If false, this identifier may or may
 	// not have side effects when referenced. This is used to allow the removal